Flesh out first iteration of database schema
Session cookies are being used to keep clients honest about their session.
The database schema in a seperate file makes working outside python easier.
It is important that the PRIMARY KEY of each table be of type INTERGER.
Otherwise, sqlite3 fails to recognize that it is a PRIMARY KEY and assigns NULL.

Without 'PRAGMA foreign_keys = 1' sqlite3 does not enforce foreign_keys constraints.
ON DELETE RESTRICT should keep the database valid even if an attempt is made to delete the referenced parent table row.
ON UPDATE CASCADE should update all the references to a row's primary key if the parent table is changed.